* partner/android-gs-pixel-mainline:
  mali_kbase: dvfs: Implement clock down hysteresis as delayed work
  mali_kbase: pm: Calculate metrics only when initialized
  Set mali_jd_thread to RT-prio 60.
  mali_mgm: Enable warnings on build
  mali_mgm: Fix build warnings
  mali_kbase: Enable warnings on build
  mali_kbase: platform: Fix Pixel integration  warnings
  mali_kbase: Change default Make target from 'all' to 'modules'
  mali_mgm: Change default Make target from 'all' to 'modules'

Bug: 169348040
Bug: 169955425
Bug: 170311684
Bug: 170498705
Bug: 171194646
Bug: 171680119

+++ mali_mgm/memory_group_manager.c
@@ -70,6 +70,7 @@ static inline vm_fault_t vmf_insert_pfn_prot(struct vm_area_struct *vma,
  * @lp_size:  The number of allocated large(2MB) pages
  * @insert_pfn: The number of calls to map pages for CPU access.
  * @update_gpu_pte: The number of calls to update GPU page table entries.
+ * @ptid: The partition ID for this group
  * @pbha: The PBHA bits assigned to this group,
  * @state: The lifecycle state of the partition associated with this group
  * This structure allows page allocation information to be displayed via
@@ -497,7 +498,7 @@ static void mgm_resize_callback(void *data, int id, size_t size_allocated)
 {
 	/* Currently we don't do anything on partition resize */
 	struct mgm_groups *const mgm_data = (struct mgm_groups *)data;
-	dev_dbg(mgm_data->dev, "Resize callback called, size_allocated: %d\n",
+	dev_dbg(mgm_data->dev, "Resize callback called, size_allocated: %zu\n",
 		size_allocated);
 }
 
@@ -552,11 +553,11 @@ static void mgm_term_data(struct mgm_groups *data)
 		if (atomic_read(&group->size) != 0)
 			dev_warn(data->dev,
 				"%zu 0-order pages in group(%d) leaked\n",
-				atomic_read(&group->size), i);
+				(size_t)atomic_read(&group->size), i);
 		if (atomic_read(&group->lp_size) != 0)
 			dev_warn(data->dev,
 				"%zu 9 order pages in group(%d) leaked\n",
-				atomic_read(&group->lp_size), i);
+				(size_t)atomic_read(&group->lp_size), i);
 
 		/* Disable partition indices and free the partition */
 		switch (group->state) {
